change: - fetch lookup to be removed - prepare useLookupPrice hook and calling whenever needed instead of calling on bridge page repeatedly.
